# Conflicts:
#	config/seata-config-core/src/main/java/io/seata/config/ConfigurationFactory.java
#	pom.xml
#	rm-datasource/src/main/java/io/seata/rm/datasource/DataSourceManager.java
#	rm-datasource/src/main/java/io/seata/rm/datasource/exec/BaseTransactionalExecutor.java
#	rm-datasource/src/main/java/io/seata/rm/datasource/exec/InsertExecutor.java
#	rm-datasource/src/main/java/io/seata/rm/datasource/exec/SelectForUpdateExecutor.java
#	rm-datasource/src/main/java/io/seata/rm/datasource/sql/druid/oracle/OracleDeleteRecognizer.java
#	rm-datasource/src/main/java/io/seata/rm/datasource/undo/AbstractUndoExecutor.java
#	rm-datasource/src/main/java/io/seata/rm/datasource/undo/AbstractUndoLogManager.java
#	rm-datasource/src/main/java/io/seata/rm/datasource/undo/oracle/OracleUndoLogManager.java
#	rm-datasource/src/main/java/io/seata/rm/datasource/undo/parser/JacksonUndoLogParser.java
#	script/config-center/config.txt
#	server/src/main/resources/file.conf

+++ core/src/main/java/io/seata/core/rpc/ChannelManager.java
@@ -38,8 +38,7 @@ import java.util.concurrent.ConcurrentMap;
 /**
  * The type channel manager.
  *
- * @author jimin.jm @alibaba-inc.com
- * @date 2018 /12/07
+ * @author slievrly
  */
 public class ChannelManager {
 
@@ -292,7 +291,7 @@ public class ChannelManager {
     private static Channel getChannelFromSameClientMap(Map<Integer, RpcContext> clientChannelMap, int exclusivePort) {
         if (null != clientChannelMap && !clientChannelMap.isEmpty()) {
             for (ConcurrentMap.Entry<Integer, RpcContext> entry : clientChannelMap.entrySet()) {
-                if (entry.getKey().intValue() == exclusivePort) {
+                if (entry.getKey() == exclusivePort) {
                     clientChannelMap.remove(entry.getKey());
                     continue;
                 }
